Add tokens for iOS
